fix(backend): dinero como Decimal en models + saldos en Decimal (#38)
- Models: Mapped[float] -> Mapped[Decimal] en los 10 campos de dinero
  (Transaction, TransactionSplit, ReconciliationSession, Account,
  RecurringRule, Budget, SavingsGoal, InstalmentPlan). Las columnas ya
  son Numeric(19,4); la anotación era incorrecta y el ORM ya devuelve
  Decimal en runtime. Sin migración.
- account_balances: balance_sums y account_balance acumulan en Decimal
  (exacto, sin artefactos binarios 0.1+0.2). Un único cast a float al
  final mantiene el contrato float de todos los consumidores.
- Tests: test_account_balances.py pinne la exactitud de Decimal en la
  fórmula de saldos (0.1+0.2+0.3 exacto).

Suite: 287 passed, 13 skipped (era 284). No se cambian schemas ni
frontend: Pydantic v2 serializa Decimal a number en JSON, así que el
contrato API no cambia.

+++ backend/app/models/accounts.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 
 from sqlalchemy import DateTime, ForeignKey, ForeignKeyConstraint, Numeric, String, UniqueConstraint, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
@@ -21,7 +22,7 @@ class Account(Base):
     owner_id: Mapped[str | None] = mapped_column(String(32), nullable=True, index=True)
     name: Mapped[str] = mapped_column(String(120))
     kind: Mapped[str] = mapped_column(String(20))
-    opening_balance: Mapped[float] = mapped_column(Numeric(19, 4), default=0)
+    opening_balance: Mapped[Decimal] = mapped_column(Numeric(19, 4), default=0)
     bank: Mapped[str | None] = mapped_column(String(60), nullable=True)
     card_brand: Mapped[str | None] = mapped_column(String(10), nullable=True)
     last_four: Mapped[str | None] = mapped_column(String(4), nullable=True)
